refactor: consolidate waitForDiagnostics into shared utility
Move diagnostic waiting logic to src/common/vscodeDiagnostics.ts so it
can be shared between Lean tools and LaTeX linter instead of having
duplicate implementations.
